requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
Babel==2.15.0
bangla==0.0.2
blinker==1.8.2
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
catalogue==2.0.10
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
cycler==0.12.1
cymem==2.0.8
Cython==3.0.10
dateparser==1.1.8
decorator==5.1.1
distro==1.9.0
docopt==0.6.2
einops==0.8.0
encodec==0.1.1
exceptiongroup==1.2.2
filelock==3.15.4
Flask==3.0.3
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
g2pkk==0.1.2
grpcio==1.65.4
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.1
gruut-lang-en==2.0.1
gruut-lang-es==2.0.1
gruut-lang-fr==2.0.2
gTTS==2.5.2
h11==0.14.0
hangul-romanize==0.1.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
inflect==7.3.1
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.4
joblib==1.4.2
jsonlines==1.2.0
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.4
librosa==0.10.0
llvmlite==0.43.0
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
murmurhash==1.0.10
networkx==2.8.8
nltk==3.8.1
num2words==0.5.13
numba==0.60.0
numpy==1.22.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==1.5.3
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
preshed==3.0.9
protobuf==4.25.4
psutil==6.0.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pynndescent==0.5.13
pyparsing==3.1.2
pypinyin==0.51.0
pysbd==0.3.4
pytesseract==0.3.10
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.11.4
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
soundfile==0.12.1
soxr==0.4.0
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
SudachiDict-core==20240716
SudachiPy==0.6.8
sympy==1.13.1
tensorboard==2.17.0
tensorboard-data-server==0.7.2
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.0
torchaudio==2.4.0
tqdm==4.66.4
trainer==0.0.36
transformers==4.43.3
triton==3.0.0
TTS==0.22.0
typeguard==4.3.0
typer==0.12.3
typing_extensions==4.12.2
tzlocal==5.2
umap-learn==0.5.6
Unidecode==1.3.8
urllib3==2.2.2
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
